refactor: extract raml specific logic from webapi reference handler
