+++ amf-webapi/shared/src/main/scala/amf/plugins/document/webapi/RamlPlugin.scala
@@ -30,6 +30,7 @@ import amf.plugins.document.webapi.parser.RamlHeader.{Raml10, Raml10Extension, R
 import amf.plugins.document.webapi.parser.spec.raml.{RamlDocumentEmitter, RamlFragmentEmitter, RamlModuleEmitter, _}
 import amf.plugins.document.webapi.parser.spec.{RamlWebApiDeclarations, WebApiDeclarations}
 import amf.plugins.document.webapi.parser.{RamlFragment, RamlHeader}
+import amf.plugins.document.webapi.references.RamlReferenceHandler
 import amf.plugins.document.webapi.resolution.pipelines.compatibility.CompatibilityPipeline
 import amf.plugins.document.webapi.resolution.pipelines.{
   Raml08EditingPipeline,
@@ -45,6 +46,8 @@ import org.yaml.model.{YDocument, YNode}
 sealed trait RamlPlugin extends BaseWebApiPlugin {
 
   override val vendors: Seq[String] = Seq(vendor.name, Raml.name)
+
+  override def referenceHandler(eh: ErrorHandler) = new RamlReferenceHandler(ID, this)
 
   def context(wrapped: ParserContext,
               root: Root,
